+++ Model/Penumbra/ImcModGroup.cs
@@ -8,7 +8,7 @@ internal class ImcModGroup : ModGroup {
     public string Name { get; set; }
     public string? Description { get; set; }
     public int Priority { get; set; }
-    public uint DefaultSettings { get; set; }
+    public ulong DefaultSettings { get; set; }
     public string Type { get; set; } = "Imc";
     public JToken Identifier { get; set; }
     public bool AllVariants { get; set; }
