+++ Model/Penumbra/ModGroup.cs
@@ -6,7 +6,7 @@ internal interface ModGroup {
     public string Name { get; set; }
     public string? Description { get; set; }
     public int Priority { get; set; }
-    public uint DefaultSettings { get; set; }
+    public ulong DefaultSettings { get; set; }
     public string Type { get; set; }
 
     [JsonIgnore]
